# api-docs.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8081","basePath":"/","tags":[{"name":"employee-controller","description":"Employee Controller"},{"name":"employee-service","description":"Employee Service"}],"paths":{"/sample/employees":{"get":{"tags":["employee-controller"],"summary":"getEmployees","operationId":"getEmployeesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Employee"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["employee-controller"],"summary":"saveEmployee","operationId":"saveEmployeeUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"employee","description":"employee","required":true,"schema":{"$ref":"#/definitions/Employee"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Employee"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/sample/employees/{empId}":{"get":{"tags":["employee-controller"],"summary":"getEmployee","operationId":"getEmployeeUsingGET","produces":["*/*"],"parameters":[{"name":"empId","in":"path","description":"empId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Employee"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["employee-controller"],"summary":"updateEmployee","operationId":"updateEmployeeUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"empId","in":"path","description":"empId","required":true,"type":"integer","format":"int32"},{"in":"body","name":"employee","description":"employee","required":true,"schema":{"$ref":"#/definitions/Employee"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Employee"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["employee-controller"],"summary":"deleteEmployee","operationId":"deleteEmployeeUsingDELETE","produces":["*/*"],"parameters":[{"name":"empId","in":"path","description":"empId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/sample/employeeservice/{empId}/salary":{"get":{"tags":["employee-service"],"summary":"getSalary","operationId":"getSalaryUsingGET","produces":["*/*"],"parameters":[{"name":"empId","in":"path","description":"empId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"number","format":"double"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"Employee":{"type":"object","properties":{"address":{"type":"string"},"designation":{"type":"string"},"empId":{"type":"integer","format":"int32"},"name":{"type":"string"}},"title":"Employee"}}}